fix: drivers: mci: Fix the MCI base address size
Fix the data returned by readl() and used as MCI indirect
configuration base address after left shift.
The code was shifting the 32-bit unsigned value before
casting it into 64-bit unsigned base (up to 40 bit address
should be supported).

Reviewed-on: http://vgitil04.il.marvell.com:8080/45935
Tested-by: iSoC Platform CI <ykjenk@marvell.com>

+++ drivers/misc/mvebu_mci.c
@@ -318,7 +318,8 @@ static int mci_probe(struct udevice *dev)
 		return PTR_ERR(mci_base_reg);
 
 	if (fdtdec_get_bool(blob, node, "indirect_base_reg"))
-		mci_base = readl(mci_base_reg) << MVEBU_MCI_PHY_BASE_OFF;
+		mci_base = (unsigned long)readl(mci_base_reg)
+						<< MVEBU_MCI_PHY_BASE_OFF;
 	else
 		mci_base = (unsigned long)mci_base_reg;
